refactor app.rb and main.rb to fullfill single responsiblity principle
